added 7000 first names to brute-force capitalize.
These names were collected using data provided to the public by the Social Security Administration. It includes 7000 of the most common names from 2000-2020, with names *manually* removed that are used in other parts of speech. (Like "Blossom", "Cliff", "Baker", etc.)

Because it was done manually, it will likely need to be updated and revisited in the future. I would prefer training the NLP to recognize these names without needing to be told, but for the time being this method works and hopefully won't raise false flags.